More renaming of files and updating unit tests
